Replaced ElevatedButtons with custom ExpandedButton
