+++ lib/widgets/padded_row_buttons.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../constants/colours.dart';
+import './expanded_button.dart';
 
 class PaddedRowButtons extends StatelessWidget {
   const PaddedRowButtons({
@@ -27,27 +27,6 @@ class PaddedRowButtons extends StatelessWidget {
             text: 'Close',
           ),
         ],
-      ),
-    );
-  }
-}
-
-class ExpandedButton extends StatelessWidget {
-  final String text;
-  final Function function;
-
-  const ExpandedButton({Key? key, required this.text, required this.function})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      child: ElevatedButton(
-        onPressed: () => function(),
-        style: ElevatedButton.styleFrom(
-          backgroundColor: Colours.kTeal,
-        ),
-        child: Text(text),
       ),
     );
   }
